Moving the test to the right place
